chore: try to fix db migrations

+++ app/migrations/versions/20260313_000000_06fc68e254b7_add_last_pool_purchase_to_teams.py
@@ -25,4 +25,8 @@ def upgrade() -> None:
 
 
 def downgrade() -> None:
-    op.drop_column("teams", "last_pool_purchase")
+    bind = op.get_bind()
+    insp = sa.inspect(bind)
+    cols = insp.get_columns("teams")
+    if any(col["name"] == "last_pool_purchase" for col in cols):
+        op.drop_column("teams", "last_pool_purchase")
